revert changes to Arrow logging
